refactor(tui): extract commitAbove phases into sibling modules (#827)
