+++ .filesize-baseline.json
@@ -50,8 +50,8 @@
     "src/agent/tools/readonly-bash.ts": { "loc": 763, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/tools/schemas.ts": { "loc": 1332, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/tools/skill-bridge.ts": { "loc": 498, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/agent/tools/skill-executor.ts": { "loc": 445, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/agent/tools/skill-executor/fork-dispatch.ts": { "loc": 384, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/agent/tools/skill-executor.ts": { "loc": 448, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/agent/tools/skill-executor/fork-dispatch.ts": { "loc": 393, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/tools/subagent-executor.ts": { "loc": 873, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/tools/subagent/child-config.ts": { "loc": 480, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/tools/subagent/foreground-promotion.ts": { "loc": 452, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
@@ -108,7 +108,6 @@
     "src/cli/slash/marketplace-browse.ts": { "loc": 367, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/slash/plugin-skills/listing.ts": { "loc": 410, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/status-line.ts": { "loc": 621, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/cli/terminal-compositor.committed-band-commit.ts": { "loc": 847, "reason": "CANDIDATE PERMANENT EXEMPTION, not yet granted: commitAbove is one atomic 3-phase terminal-escape transaction (scrollback archive -> live repaint -> committed paint) whose inline invariant comments are the only guard against regressing #81/#403/#467, and scattering its intermediate values across a file boundary would convert compiler-invisible sequential correctness into compiler-invisible cross-file correctness. Grew 474 -> 847 since 2026-06-10 across 16 commits. A wave must attempt the split and prove infeasibility before this is promoted to permanent:true." },
     "src/cli/terminal-compositor.frame.ts": { "loc": 370, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/terminal-compositor.input-dispatch.ts": { "loc": 1153, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/terminal-compositor.lifecycle.ts": { "loc": 483, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
