+++ .funcsize-baseline.json
@@ -6,9 +6,10 @@
     "src/agent/hooks/command-executor.ts::executeCommand": { "loc": 237, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/anthropic-direct/loop.ts::runTurn": { "loc": 245, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/anthropic-direct/loop/stream-consumer.ts::consumeRoundStream": { "loc": 279, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/agent/providers/anthropic-direct/query-turn-driver.ts::driveTurns": { "loc": 224, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/anthropic-direct/translate.ts::translateMessageStream": { "loc": 209, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/openai-compatible/index.ts::OpenAICompatibleProvider.query": { "loc": 240, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/agent/providers/openai-compatible/query.ts::OpenAICompatibleQuery._runTurnInner": { "loc": 277, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/agent/providers/openai-compatible/query.ts::OpenAICompatibleQuery._runTurnInner": { "loc": 297, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/providers/openai-compatible/query/dispatch-append.ts::dispatchAndAppendToolCalls": { "loc": 233, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/session/stream-consumer.ts::transformProviderEvent": { "loc": 241, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/agent/shell-jobs/streamer.ts::startShell": { "loc": 290, "reason": "legacy: predates the function-size gate; pending helper extraction" },
@@ -26,14 +27,14 @@
     "src/cli/_lib/stream-renderer-orchestrator.ts::handleOrchestratorEvent": { "loc": 304, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/_lib/stream-renderer-subagent.ts::handleSubagentEvent": { "loc": 254, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/_lib/stream-renderer.ts::StreamRenderer.process": { "loc": 203, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/chat.ts::registerChatCommand": { "loc": 678, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/chat.ts::registerChatCommand": { "loc": 683, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/config-command.ts::registerConfigCommand": { "loc": 204, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/daemon.ts::registerDaemonCommand": { "loc": 293, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/farm.ts::runFarm": { "loc": 243, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive.ts::registerInteractiveCommand": { "loc": 713, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/interactive/bootstrap.ts::bootstrapSession": { "loc": 261, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/interactive/bootstrap.ts::bootstrapSession": { "loc": 319, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/footer-subsystems.ts::setupFooterSubsystems": { "loc": 211, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/interactive/loop-iteration.ts::runInputLoop": { "loc": 713, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/interactive/loop-iteration.ts::runInputLoop": { "loc": 711, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/surface-setup.ts::setupSurface": { "loc": 302, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/tool-lane-render-children.ts::renderOverlayChildren": { "loc": 245, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/tool-lane.ts::ToolLane.getOverlay": { "loc": 229, "reason": "legacy: predates the function-size gate; pending helper extraction" },
@@ -45,14 +46,13 @@
     "src/cli/elicitation/agent-question.ts::renderAgentQuestion": { "loc": 400, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/formatter.ts::renderMarkdownToTerminal": { "loc": 405, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/input/reader.ts::readWithAutocompleteTty": { "loc": 856, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/terminal-compositor.committed-band-commit.ts::commitAbove": { "loc": 735, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/terminal-compositor.frame-preserve.ts::preserveRowsBeforeFrameRender": { "loc": 209, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/terminal-compositor.input-dispatch.ts::handleEnter": { "loc": 259, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/terminal-compositor.lifecycle.ts::arm": { "loc": 206, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/skills/audit-fit/index.ts::handler": { "loc": 291, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/telegram/bot.ts::TelegramBot.setupHandlers": { "loc": 241, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/telegram/elicitation-handler.ts::makeTelegramElicitationHandler": { "loc": 371, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/telegram/handlers/message.ts::MessageHandler.handlePhoto": { "loc": 232, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/telegram/bot.ts::TelegramBot.setupHandlers": { "loc": 247, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/telegram/elicitation-handler.ts::makeTelegramElicitationHandler": { "loc": 381, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/telegram/handlers/message.ts::MessageHandler.handlePhoto": { "loc": 234, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/telegram/streaming.ts::streamResponse": { "loc": 769, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/web-server/frontend/view-model.ts::createTranscriptModel": { "loc": 208, "reason": "legacy: predates the function-size gate; pending helper extraction" }
   }
